third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/event-listener-breakpoints-script-first-stmt.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    setTimeout(callback1, 0);
}

function callback1()
{
    var code =
        "(function() {\n" +
            "setTimeout(callback2, 0);\n" +
        "})();\n" +
        "//# sourceURL=inline_callback1.js";
    setTimeout(code, 0);
}

function callback2()
{
    var script = document.createElement("script");
    script.src = "../debugger/resources/script1.js";
    document.body.appendChild(script);
    script.remove();
}

function test() {
  var numberOfPauses = 2;

  SourcesTestRunner.startDebuggerTest(step1, true);

  function step1() {
    SourcesTestRunner.setEventListenerBreakpoint('instrumentation:scriptFirstStatement', true);
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(didPause);
  }

  function didPause(callFrames, reason, breakpointIds, asyncStackTrace, auxData) {
    var eventName = (auxData && auxData.eventName || '').replace(/^instrumentation:/, '');
    TestRunner.addResult('\nPaused on ' + eventName);
    SourcesTestRunner.captureStackTrace(callFrames);

    if (--numberOfPauses)
      SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(InspectorTest, didPause));
    else
      SourcesTestRunner.resumeExecution(step2);
  }

  function step2() {
    SourcesTestRunner.setEventListenerBreakpoint('instrumentation:scriptFirstStatement', false);
    SourcesTestRunner.completeDebuggerTest();
  }
}

</script>
</head>

<body onload="runTest()">
<input type='button' onclick='testFunction()' value='Test'/>
<p>
Tests event listener breakpoint to break on the first statement of new scripts.
</p>
</body>
</html>
