third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/event-listener-breakpoints-script-first-stmt-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests event listener breakpoint to break on the first statement of new scripts.

Set timer for test function.

Paused on scriptFirstStatement
Call stack:
    0)  (inline_callback1.js:1)

Paused on scriptFirstStatement
Call stack:
    0)  (script1.js:1)
